+++ packages/date/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2](https://github.com/Availity/availity-react/compare/@availity/date@0.3.1...@availity/date@0.3.2) (2019-10-01)
+
+**Note:** Version bump only for package @availity/date
+
+
+
+
+
+## [0.3.1](https://github.com/Availity/availity-react/compare/@availity/date@0.3.0...@availity/date@0.3.1) (2019-10-01)
+
+
+### Bug Fixes
+
+* **DateField:** use id as for in label ([2d976b9](https://github.com/Availity/availity-react/commit/2d976b9))
+* **DateRangeField:** make label target start field ([1099bf2](https://github.com/Availity/availity-react/commit/1099bf2))
+
+
+
+
+
+# [0.3.0](https://github.com/Availity/availity-react/compare/@availity/date@0.2.9...@availity/date@0.3.0) (2019-10-01)
+
+
+### Features
+
+* **date:** added additional attributes to the label for DateField and DateRangeField ([f321dcd](https://github.com/Availity/availity-react/commit/f321dcd))
+* **date:** added new props to type definitions ([95d9aa4](https://github.com/Availity/availity-react/commit/95d9aa4))
+* **date:** added new props to type definitions ([867dfb1](https://github.com/Availity/availity-react/commit/867dfb1))
+
+
+
+
+
+## [0.2.9](https://github.com/Availity/availity-react/compare/@availity/date@0.2.8...@availity/date@0.2.9) (2019-09-26)
+
+**Note:** Version bump only for package @availity/date
+
+
+
+
+
+## [0.2.8](https://github.com/Availity/availity-react/compare/@availity/date@0.2.7...@availity/date@0.2.8) (2019-09-25)
+
+**Note:** Version bump only for package @availity/date
+
+
+
+
+
+## [0.2.7](https://github.com/Availity/availity-react/compare/@availity/date@0.2.6...@availity/date@0.2.7) (2019-09-23)
+
+**Note:** Version bump only for package @availity/date
+
+
+
+
+
 ## [0.2.6](https://github.com/Availity/availity-react/compare/@availity/date@0.2.5...@availity/date@0.2.6) (2019-09-19)
 
 **Note:** Version bump only for package @availity/date