+++ packages/favorites/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.2](https://github.com/Availity/availity-react/compare/@availity/favorites@2.1.1...@availity/favorites@2.1.2) (2019-09-26)
+
+**Note:** Version bump only for package @availity/favorites
+
+
+
+
+
+## [2.1.1](https://github.com/Availity/availity-react/compare/@availity/favorites@2.1.0...@availity/favorites@2.1.1) (2019-09-25)
+
+**Note:** Version bump only for package @availity/favorites
+
+
+
+
+
 # [2.1.0](https://github.com/Availity/availity-react/compare/@availity/favorites@2.0.1...@availity/favorites@2.1.0) (2019-09-04)