+++ packages/feedback/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.1.11](https://github.com/Availity/availity-react/compare/@availity/feedback@5.1.10...@availity/feedback@5.1.11) (2019-10-01)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [5.1.10](https://github.com/Availity/availity-react/compare/@availity/feedback@5.1.9...@availity/feedback@5.1.10) (2019-09-26)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [5.1.9](https://github.com/Availity/availity-react/compare/@availity/feedback@5.1.8...@availity/feedback@5.1.9) (2019-09-25)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [5.1.8](https://github.com/Availity/availity-react/compare/@availity/feedback@5.1.7...@availity/feedback@5.1.8) (2019-09-23)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
 ## [5.1.7](https://github.com/Availity/availity-react/compare/@availity/feedback@5.1.6...@availity/feedback@5.1.7) (2019-09-19)
 
 **Note:** Version bump only for package @availity/feedback