+++ packages/list-group/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/Availity/availity-react/compare/@availity/list-group@1.2.0...@availity/list-group@1.2.1) (2019-09-25)
+
+**Note:** Version bump only for package @availity/list-group
+
+
+
+
+
 # [1.2.0](https://github.com/Availity/availity-react/compare/@availity/list-group@1.1.8...@availity/list-group@1.2.0) (2019-09-04)