+++ packages/payer-logo/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.3](https://github.com/Availity/availity-react/compare/@availity/payer-logo@4.0.2...@availity/payer-logo@4.0.3) (2019-09-26)
+
+**Note:** Version bump only for package @availity/payer-logo
+
+
+
+
+
+## [4.0.2](https://github.com/Availity/availity-react/compare/@availity/payer-logo@4.0.1...@availity/payer-logo@4.0.2) (2019-09-25)
+
+**Note:** Version bump only for package @availity/payer-logo
+
+
+
+
+
 ## [4.0.1](https://github.com/Availity/availity-react/compare/@availity/payer-logo@4.0.0...@availity/payer-logo@4.0.1) (2019-09-04)
 
 **Note:** Version bump only for package @availity/payer-logo